+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 next
 ====
 
+- Bumped dependencies to pybind11≥2.6.
 - Improve snapping of codeful paths.
 - Fix failure to increase cairo refcount of font cache entries.